various development updates
