+++ soc/src/main/scala/AbstractConfig.scala
@@ -32,7 +32,6 @@ class AbstractConfig extends Config(
   new chipyard.iobinders.WithBlockDeviceIOPunchthrough ++
   new chipyard.iobinders.WithNICIOPunchthrough ++
   new chipyard.iobinders.WithAirSimIOPunchthrough ++
-  new chipyard.iobinders.WithRoseIOPunchthrough ++
   new chipyard.iobinders.WithSerialTLIOCells ++
   new chipyard.iobinders.WithDebugIOCells ++
   new chipyard.iobinders.WithUARTIOCells ++
