+++ soc/src/main/scala/RoSEBridge.scala
