+++ soc/src/main/scala/RoSEConfigs.scala
@@ -1,10 +1,9 @@
+//These are the examples configs for the RoSE-supported SoC
+// *** Not to be mistaken with the configurations for the RoSE adapters ***
 package chipyard
 
 import freechips.rocketchip.config.{Config}
 
-// ---------------------
-// BOOM Configs
-// ---------------------
 class AirSimIOTLFPGemminiLargeBoomConfig extends Config(
   new chipyard.example.WithAirSimIO(useAXI4=false) ++          // Usesim/src/main/cc/firesim/firesim_top.cc GCD Chisel, connect Tilelink
   new gemmini.GemminiFP32DefaultConfig ++                         // use FP32Gemmini systolic array GEMM accelerator
@@ -24,6 +23,10 @@ class AirSimIOTLFPGemminiRocketConfig extends Config(
   new chipyard.config.AbstractConfig)
 
 class RoseTLRocketConfig extends Config(
-  new rose.WithRoseAdapter(useAXI4=false, base=0x88000000L, width=32) ++          // Use GCD Chisel, connect Tilelink                      // use FP32Gemmini systolic array GEMM accelerator
+  new rose.WithRoseAdapter(base=0x88000000L, width=32) ++          // Use GCD Chisel, connect Tilelink                      // use FP32Gemmini systolic array GEMM accelerator
   new freechips.rocketchip.subsystem.WithNBigCores(1) ++
+  new chipyard.config.AbstractRoseConfig)
+
+class AbstractRoseConfig extends Config(
+  new chipyard.iobinders.WithRoseIOPunchthrough ++
   new chipyard.config.AbstractConfig)
